requirements.txt
requests>=2.32.0
openai>=1.6.0
python-dotenv>=1.0.0
markdown>=3.6

